store/store_notifications/requirements.txt
asgiref==3.6.0
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
filelock==3.9.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
gunicorn==20.1.0
virtualenv==20.19.0

